feat(carousel): convert to typescript
