+++ src/components/CardContainerRating.tsx
@@ -32,7 +32,7 @@ interface Props {
   elevation?: number;
   numColumns?: number;
   theme: typeof theme;
-  style: StyleProp<ViewStyle>;
+  style?: StyleProp<ViewStyle>;
   onPress: () => void;
 }
